RCM-Analyzer/requirements.txt
streamlit==1.31.0
pandas==2.1.3
numpy==1.26.2
chromadb==0.4.18
# pysqlite3-binary==0.5.1 (removed, use built-in sqlite3)
google-generativeai==0.3.1
PyPDF2==3.0.1
langchain>=0.0.344
langchain-core>=0.1.16
python-dotenv==1.0.0
openpyxl==3.1.2
plotly==5.18.0
tqdm==4.66.1
pdfplumber==0.10.3
python-docx==1.0.1
pillow==10.1.0
sentence-transformers==2.2.2 